eon and enterprise to use different set of links (#16)
Eon dashboards are now tagged with vertica-eon. Enterprise are tagged
with vertica-enterprise. This will allow the dashboard link to only
provide other dashboards for the same deployment type.

Here is a sample dashboard for enterprise (notice link drop down in the
top right):

![image](https://user-images.githubusercontent.com/171762/219168591-14fa75ec-7db6-4d05-abc8-7ab9008a9e2a.png)

Here is a sample dashboard for eon:

![image](https://user-images.githubusercontent.com/171762/219168978-27c7d4d6-5d89-4c1a-b8d4-d0c27f48ea65.png)